src/main/java/com/imooc/sellbymyself/service/impl/CategoryServiceImpl.java
package com.imooc.sellbymyself.service.impl;

import com.imooc.sellbymyself.dataobject.ProductCategory;
import com.imooc.sellbymyself.repository.ProductCategoryRepository;
import com.imooc.sellbymyself.service.CategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * 类目
 * @author hekai
 * @date 2018/9/4 0004 下午 17:00
 */
@Service
public class CategoryServiceImpl implements CategoryService {

    @Autowired
    private ProductCategoryRepository repository;

    @Override
    public ProductCategory findOne(Integer categoryId) {
        return repository.findOne(categoryId);
    }

    @Override
    public List<ProductCategory> findAll() {
        return repository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryTypeList) {
        return repository.findByCategoryTypeIn(categoryTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return repository.save(productCategory);
    }
}
